requirements.txt
install==1.3.5
Pillow==9.5.0
PySimpleGUI==4.60.4
